Move backends to seperate packages
This moves the backends each to its own package, so that users do not
have to include support for every backend (and the somewhat large
dependencies associated with them) even if they are not needed.

To enable a backend for url based usage, simply do an anonymous import,
just as you would with a sql driver.

Note that this represents a breaking API change.

Fixes #20
Updates #21 and #22
